refactor: split RangeFilterInput into separate number and date components
Extracts shared utilities into rangeFilterUtils.ts, and moves number and
date logic into NumberRangeFilterInput and DateRangeFilterInput respectively.
RangeFilterInput becomes a thin dispatcher.
